fix #121490
